* Commit patch to correctly send unicode strings on the wire (Trac #47 -- thanks monaka!) * Change -[CRDServerList rightMouseDown:] to perform own contextual menu pop-up, avoiding contextual menu popup for labels. * Change -[CRDServerList selectRow:] to use NSMutableIndexSet for private ivar _selectedRows, like the 10.5 NSTableView.h calls for

+++ Source/AppController.m
@@ -1348,7 +1348,7 @@
 	return YES;
 }
 
-- (BOOL)tableView:(NSTableView *)aTableView canDragRow:(unsigned)rowIndex
+- (BOOL)tableView:(NSTableView *)aTableView canDragRow:(NSUInteger)rowIndex
 {	
 	if (_isFilteringSavedServers)
 		return [filteredServers indexOfObject:[self serverInstanceForRow:rowIndex]] != NSNotFound;
@@ -1356,7 +1356,7 @@
 		return [savedServers indexOfObject:[self serverInstanceForRow:rowIndex]] != NSNotFound;
 }
 
-- (BOOL)tableView:(NSTableView *)aTableView canDropAboveRow:(unsigned)rowIndex
+- (BOOL)tableView:(NSTableView *)aTableView canDropAboveRow:(NSUInteger)rowIndex
 {
 	if (_isFilteringSavedServers)
 		return NO;
